# backend/docs/openapi.v0.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "duoduo-product-testing-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/personas": {
      "get": {
        "tags": [
          "personas"
        ],
        "summary": "List Personas",
        "description": "List personas visible to the current user.",
        "operationId": "list_personas_api_v1_personas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "include_critical",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Critical"
            }
          },
          {
            "name": "owner_scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "system",
                "mine",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Owner Scope"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "personas"
        ],
        "summary": "Create Persona",
        "description": "Create a private persona owned by the current user.",
        "operationId": "create_persona_api_v1_personas_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/personas/recommend": {
      "get": {
        "tags": [
          "personas"
        ],
        "summary": "Recommend Personas",
        "description": "Recommend personas for a product.",
        "operationId": "recommend_personas_api_v1_personas_recommend_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/personas/{persona_id}": {
      "get": {
        "tags": [
          "personas"
        ],
        "summary": "Get Persona",
        "description": "Return one full persona visible to the current user.",
        "operationId": "get_persona_api_v1_personas__persona_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "personas"
        ],
        "summary": "Update Persona",
        "description": "Update a private persona owned by the current user.",
        "operationId": "update_persona_api_v1_personas__persona_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Persona Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "personas"
        ],
        "summary": "Delete Persona",
        "description": "Soft-delete a private persona owned by the current user.",
        "operationId": "delete_persona_api_v1_personas__persona_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evaluations/{evaluation_id}/personas": {
      "put": {
        "tags": [
          "evaluations"
        ],
        "summary": "Update Evaluation Personas",
        "description": "Update selected personas for an editable evaluation.",
        "operationId": "update_evaluation_personas_api_v1_evaluations__evaluation_id__personas_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evaluation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationPersonaUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluationPersonaUpdateRequest": {
        "properties": {
          "persona_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Persona Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "persona_ids"
        ],
        "title": "EvaluationPersonaUpdateRequest",
        "description": "Selected personas update request."
      },
      "EvaluationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "survey_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Survey Id"
          },
          "selected_persona_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Persona Ids"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "credit_cost": {
            "type": "integer",
            "title": "Credit Cost"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationStats"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "user_id",
          "product_id",
          "survey_id",
          "selected_persona_ids",
          "status",
          "progress",
          "credit_cost",
          "created_at"
        ],
        "title": "EvaluationResponse",
        "description": "Evaluation response aligned to the API contract."
      },
...
      "EvaluationStats": {
        "properties": {
          "total_personas": {
            "type": "integer",
            "title": "Total Personas"
          },
          "completed_personas": {
            "type": "integer",
            "title": "Completed Personas"
          },
          "failed_personas": {
            "type": "integer",
            "title": "Failed Personas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total_personas",
          "completed_personas",
          "failed_personas"
        ],
        "title": "EvaluationStats",
        "description": "Evaluation progress stats."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OceanScores": {
        "properties": {
          "o": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "O"
          },
          "c": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "C"
          },
          "e": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "E"
          },
          "a": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "A"
          },
          "n": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "N"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "o",
          "c",
          "e",
          "a",
          "n"
        ],
        "title": "OceanScores",
        "description": "Big Five personality scores from the public API."
      },
...
      "PersonaCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "age": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Age"
          },
          "gender": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Gender"
          },
          "city": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "City"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "income_monthly": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Income Monthly"
          },
          "persona_tag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Tag"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Categories"
          },
          "is_critical": {
            "type": "boolean",
            "title": "Is Critical",
            "default": false
          },
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile"
          },
          "ocean": {
            "$ref": "#/components/schemas/OceanScores"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "age",
          "gender",
          "city",
          "categories",
          "profile",
          "ocean"
        ],
        "title": "PersonaCreateRequest",
        "description": "Private persona creation request."
      },
      "PersonaListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "city_tier": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Tier"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "income_monthly": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Income Monthly"
          },
          "persona_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Tag"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "is_critical": {
            "type": "boolean",
            "title": "Is Critical"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "avatar",
          "age",
          "gender",
          "city",
          "city_tier",
          "occupation",
          "income_monthly",
          "persona_tag",
          "categories",
          "is_critical",
          "is_system"
        ],
        "title": "PersonaListItem",
        "description": "Simplified persona item for lists and recommendations."
      },
      "PersonaPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonaListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PersonaPageResponse",
        "description": "Offset-paginated persona response."
      },
      "PersonaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "city_tier": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Tier"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "income_monthly": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Income Monthly"
          },
          "persona_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Tag"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "is_critical": {
            "type": "boolean",
            "title": "Is Critical"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System"
          },
          "ocean": {
            "$ref": "#/components/schemas/OceanScores"
          },
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "avatar",
          "age",
          "gender",
          "city",
          "city_tier",
          "occupation",
          "income_monthly",
          "persona_tag",
          "categories",
          "is_critical",
          "is_system",
          "ocean",
          "profile",
          "version",
          "created_at"
        ],
        "title": "PersonaResponse",
        "description": "Full persona response."
      },
...
      "PersonaUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "income_monthly": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Income Monthly"
          },
          "persona_tag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Tag"
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 20,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories"
          },
          "is_critical": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Critical"
          },
          "profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "ocean": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OceanScores"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PersonaUpdateRequest",
        "description": "Private persona partial update request."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}